py/requirements.txt
Markdown==3.4.4
requests==2.32.3
python-dotenv==1.0.1
requests-oauthlib==2.0.0
polars==1.9.0
autopep8==2.3.1
flake8==7.1.1
fastexcel==0.12.0
numpy==2.1.2
XlsxWriter==3.2.0